Implement uncategorized posts
If a post appears in the root directory, it becomes an uncategorized post.
An uncategorized post belongs to the category 'Uncategorized'.

It's a dynamic category that appears if there is an uncategorized post in the site.

Fix a bug that happens when a file with space char in its name is in the root dir. Use escape characters to feed the file name to cp command.

Minor improvements